rewrite MetricCollector by NodeCollector but failed.
This looks so weird.

I have to manually call the fetch function to get new metric on hand.
This will couple the responsiblity of metric fetching & metric
operation.

Of course I can schedule this simple operation to somewhere else, maybe
with the help of schedule executor. But to get metric I have to call
that `Receiver#current()` function, which might cause a unfortunate
loading of metric. This might not be what I want. So yeah no.